% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-data.R
\name{fi_data}
\alias{fi_data}
\title{Post data to file.io}
\usage{
fi_data(text, expires = "14d")
}
\arguments{
\item{text}{single element character vector}

\item{expires}{defaults to 14 days (\code{14d}). Must be a positive integer which,
by default, represents the number of days until the file will be deleted.
If you follow it with \code{w}, it will be the number of weeks; \code{m} for months;
and, \code{y} for years.}
}
\value{
a data frame (tibble) with the success status, download key, download link
and expiry
}
\description{
Post data to file.io
}
\examples{
fi_data("Hi Noam!")
}
\references{
\url{https://www.file.io/}
}
