% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-file.R
\name{fi_post}
\alias{fi_post}
\title{Post a file!}
\usage{
fi_post(path, expires = "14d")
}
\arguments{
\item{path}{path to a file}

\item{expires}{defaults to 14 days (\code{14d}). Must be a positive integer which,
by default, represents the number of days until the file will be deleted.
If you follow it with \code{w}, it will be the number of weeks; \code{m} for months;
and, \code{y} for years.}
}
\value{
a data frame (tibble) with the success status, download key, download link
and expiry
}
\description{
Post a file!
}
\examples{
fi_post(system.file("extdat", "tst.txt", package = "fileio"))
}
