% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-file.R
\name{fi_post_file}
\alias{fi_post_file}
\title{Post a file to file.io}
\usage{
fi_post_file(path, expires = "14d")
}
\arguments{
\item{path}{path to a file (it will be \code{\link[=path.expand]{path.expand()}}ed and \code{\link[=normalizePath]{normalizePath()}}'d)}

\item{expires}{defaults to 14 days (\code{14d}). Must be a positive integer which,
by default, represents the number of days until the file will be deleted.
If you follow it with \code{w}, it will be the number of weeks; \code{m} for months;
and, \code{y} for years.}
}
\value{
a data frame (tibble) with the success status, download key, download link
and expiry
}
\description{
Post a file to file.io
}
\note{
There is a 5GB per file limit for the free version.
}
\examples{
fi_post_file(system.file("extdat", "tst.txt", package = "fileio"))
}
\references{
\url{https://www.file.io/}
}
