% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-data.R
\name{fi_post_rdata}
\alias{fi_post_rdata}
\title{Post an R object to file.io}
\usage{
fi_post_rdata(robj, filename = uuid::UUIDgenerate(), expires = "14d")
}
\arguments{
\item{robj}{an R object}

\item{filename}{if not provided \code{\link[uuid:UUIDgenerate]{uuid::UUIDgenerate()}} will be used to
create a file and that will be returned in the result data frame}

\item{expires}{defaults to 14 days (\code{14d}). Must be a positive integer which,
by default, represents the number of days until the file will be deleted.
If you follow it with \code{w}, it will be the number of weeks; \code{m} for months;
and, \code{y} for years.}
}
\value{
a data frame (tibble) with the success status, download key, download link
and expiry
}
\description{
This will take an R object (which can be a list of serializable objects) and
create an RDS file from it and send it off to file.io.
}
\note{
There is a 5GB per file limit for the free version.
}
\examples{
fi_post_rdata(
  list(
    mtcars = mtcars,
    iris = iris,
    message = "Hi Noam!"
  )
)
}
\references{
\url{https://www.file.io/}
}
