% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-text.R
\name{fi_post_text}
\alias{fi_post_text}
\alias{fi_text}
\title{Post text to file.io}
\usage{
fi_post_text(text, expires = "14d")

fi_text(text, expires = "14d")
}
\arguments{
\item{text}{single element character vector}

\item{expires}{defaults to 14 days (\code{14d}). Must be a positive integer which,
by default, represents the number of days until the file will be deleted.
If you follow it with \code{w}, it will be the number of weeks; \code{m} for months;
and, \code{y} for years.}
}
\value{
a data frame (tibble) with the success status, download key, download link
and expiry
}
\description{
Post text to file.io
}
\note{
There is a 5GB per file limit for the free version.
}
\examples{
x <- fi_post_text("Hi Noam!", expires = "1d")
con <- url(x$link)
readLines(con, warn = FALSE)
close(con)
}
\references{
\url{https://www.file.io/}
}
