% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileio-package.R
\docType{package}
\name{fileio}
\alias{fileio}
\alias{fileio-package}
\title{Post Files, Text or R Data to 'file.io'}
\description{
The 'file.io' <file.io> service enables ephemeral, convenient
and anonymous file sharing. Methods are provided to upload existing files,
R data objects or text messages to this service.
}
\references{
\href{https://www.file.io/privacy.html}{file.io privacy policy}
}
\author{
Bob Rudis (bob@rud.is)
}
