% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileio-package.R
\docType{package}
\name{fileio}
\alias{fileio}
\alias{fileio-package}
\title{Post Files, Text or R Data to 'file.io'}
\description{
Post Files, Text or R Data to 'file.io'
}
\author{
Bob Rudis (bob@rud.is)
}
